backend-djangorest/api/src/pyproject.toml
[tool.poetry]
name = "Currency Conversion API"
version = "1.0.3"
description = ""
authors = ["FabboMaster"]
license = "Apache License 2.0"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.1"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"
django-configurations = "^2.4.1"
django-apscheduler = "^0.6.2"
django-filter = "^23.2"
django-cors-headers = "^4.0.0"
drf-yasg = "^1.21.5"
psycopg2-binary = "^2.9.6"
coverage = "^7.2.7"
beautifulsoup4 = "^4.12.2"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
